Add system for fetching menus
Added class for crawlers to register to (called MenuFetcher) so they can
be executed later when wanted (by server for now). Did a poor attempt
(Object.freeze) at making it singleton.

It currently imports the crawlers (just SaltoFiskhallCrawler at the
moment) so that they register themselves to the menu fetcher. It violates
the open/closed principle but I don't know how to do it better in
TypeScript at the moment.

Moved the code for crawling the saltofiskhall.se-site to its own class
which implements the new ICrawler interface. It's anonymous since it
should be executed through the menu fetcher.